Add files via upload
The Student Record Management project is designed in ‘C++’ language. This system is based on a concept to store and generate all the records of the students. Also, this program has a simple database of students for a school where the user can add student details safely and it’s not time-consuming. This System makes it easy to store records of each. Moreover, the user can view all the records, modify and remove the details. Also, this mini project is easy to operate and understand by the users.